+++ UITests/NablaUITestsHostApp/NablaUITestsHostApp/Mocks/ConversationViewControllerUITests/NablaMessagingProtocolMock+testSendMediaMessage.swift
@@ -37,7 +37,7 @@ extension NablaMessagingClientProtocolMock {
                                 fileName: "image",
                                 fileUrl: URL(string: "https://avatars.githubusercontent.com/u/39350711?s=200&v=4")!, // swiftlint:disable:this force_unwrapping
                                 thumbnailUrl: URL(string: "https://avatars.githubusercontent.com/u/39350711?s=200&v=4")!, // swiftlint:disable:this force_unwrapping
-                                mimeType: .png
+                                mimeType: .image(.png)
                             )
                         ),
                     ]